fix(arrears): stable pagination, and Exco totals from the database not the browser
Prod says the fan-out is gone: zero duplicates on all five join keys, and the
arrears book is 8,235 accounts / KSh 9.70bn arrears / KSh 21.97bn outstanding.
Nothing is being multiplied. But that ground truth exposed a different way the
Exco Loan Arrears card was reporting a wrong number, and this fixes it.

1. ExcoLoansArrearsListView had no ordering at all.

   An unordered queryset paginated with LIMIT/OFFSET gives Postgres no
   obligation to arrange rows the same way twice — Django warns about it
   (UnorderedObjectListWarning). A client that walks all 824 pages of the
   arrears book collected some loans more than once and never saw others, so a
   total summed from the crawl was wrong by a different amount on each reload.
   Now ordered by primary key.

2. The four raw-SQL arrears lists had the same defect less obviously.

   They ORDER BY days_in_arrears then a grade CASE — thousands of ties and no
   unique tiebreaker — and the views hand the result to DRF a page at a time.
   Added loan_account_no as the tiebreaker so the total order is deterministic.

3. The Exco KPI cards summed the crawled list in the browser.

   The shared page crawler stops at 500 pages of 10, so it could hold at most
   5,000 of the 8,235 accounts in arrears: those cards were structurally
   incapable of showing the real book, roughly 40% short before the ordering
   problem is even considered. New exco/loans-arrears/summary/ returns the same
   whole-bank query the CEO dashboard uses — straight off `loans`, no joins,
   nothing to truncate — and the cards read that, keeping the reduce() only as a
   fallback for when the endpoint is unavailable. The third card now reads
   "Customers in Arrears", which is what the summary actually counts.

Also fixes round(double precision, int) in docs/arrears-doubling-check.sql;
total_arrears is a float, and that overload does not exist in Postgres.

+++ docs/arrears-doubling-check.sql
@@ -115,5 +115,7 @@ unfixed AS (
 SELECT t.acc   AS true_accounts,   t.arr AS true_arrears,
        f.acc   AS fixed_accounts,  f.arr AS fixed_arrears,
        u.acc   AS unfixed_accounts,u.arr AS unfixed_arrears,
-       round(u.arr / nullif(t.arr, 0), 4) AS inflation_before_fix
+       -- cast: total_arrears is double precision, and round(double, int)
+       -- does not exist in Postgres.
+       round((u.arr / nullif(t.arr, 0))::numeric, 4) AS inflation_before_fix
 FROM truth t, fixed f, unfixed u;
